6.7 - Created AnecdoteForm.js
