+++ part6/redux-anecdotes-main/src/App.js
@@ -1,6 +1,7 @@
 import { useSelector, useDispatch } from 'react-redux'
 import { voteAction, addAction } from './reducers/anecdoteReducer'
 import './App.css'
+import AnecdoteForm from './components/AnecdoteForm'
 
 const App = () => {
   const anecdotes = useSelector(state => state)
@@ -11,20 +12,12 @@ const App = () => {
     dispatch(voteAction(id))
   }
 
-  const add = event => {
-    event.preventDefault()
-
-    const content = event.target.content.value
-    event.target.content.value = ''
-    console.log('add', content)
-    dispatch(addAction(content))
-  }
 
   return (
     <div>
       <h2>Anecdotes</h2>
       {anecdotes.map(anecdote =>
-        <div key={anecdote.id}>
+        <div className='anecdote' key={anecdote.id}>
           <div>
             {anecdote.content}
           </div>
@@ -34,11 +27,7 @@ const App = () => {
           </div>
         </div>
       )}
-      <h2>create new</h2>
-      <form onSubmit={add}>
-        <div><input name='content' /></div>
-        <button type='submit'>create</button>
-      </form>
+      <AnecdoteForm />
     </div>
   )
 }
